refactor: remove test-only seams from production, reorganize tests, close reachable coverage gaps
Production cleanup (no test-injection hooks left in shipping code):
- webauthn: drop package-level function-var seams (ioReadFull/randReader/
  uuidNewRandom/aaguidMarshalBinary/tx*Fn/getSessionByTokenHashFn/deleteSessionFn)
  and the DeleteCredential split that existed only for them.
- ratelimit: drop the cleanupInterval var; backup: drop schedulerInitialDelay
  field + backupSchedulerInitialDelay var.
- cmd/server: replace the staticFS global-swap seam with idiomatic DI — SPAHandler
  owns its fs.FS via an unexported newSPAHandler(fsys) constructor (also stops
  ServeHTTP re-deriving the sub-FS per request).

Tests:
- delete tests that only passed via those seams (rand/uuid/marshal/tx failures,
  5-min ticker, 1-min scheduler delay) and the no-op t.Log "documentation" tests.
- reorganize the coverage_targets{,2,3}_test.go junk-drawer files into the
  matching subject files.
- add tests for genuinely reachable branches: publishDiscoveryEvent severities,
  banner stdout wrappers, LoadModelsDev, GetSettings nil-map, virtual-key
  rate-limit validation, ListLogs sort_by paths, app-log cursor sort_dir,
  buildModelListQuery before-direction flip.

Frontend:
- polyfill Element.scrollIntoView in the vitest setup (jsdom gap) so the
  language-selector suite passes.
- FailoverGroups: pb-0 -> pb-6 so the last card isn't flush to the viewport.

golangci-lint clean; full backend suite and affected frontend suites green.
