--- internal/api/coverage_targets2_test.go
@@ -1,635 +1,0 @@
-package api
-
-import (
-	"bytes"
-	"context"
-	"encoding/json"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"strings"
-	"testing"
-
-	"github.com/go-chi/chi/v5"
-	"github.com/google/uuid"
-
-	"github.com/hugalafutro/model-hotel/internal/db"
-	"github.com/hugalafutro/model-hotel/internal/provider"
-)
-
-// ---------------------------------------------------------------------------
-// 1. UpdateSettings — int-type value below minimum
-// ---------------------------------------------------------------------------
-
-func TestUpdateSettings_IntBelowMin(t *testing.T) {
-	_, r := newTestHandlerWithRouter(t)
-
-	// rate_limit_burst min is 1, so 0 should fail
-	body := `{"rate_limit_burst": "0"}`
-	req := httptest.NewRequest(http.MethodPut, "/settings", strings.NewReader(body))
-	req.Header.Set("Authorization", "Bearer test-admin-token")
-	req.Header.Set("Content-Type", "application/json")
-
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusBadRequest {
-		t.Fatalf("expected status 400 Bad Request, got %d: %s", w.Code, w.Body.String())
-	}
-
-	if !strings.Contains(w.Body.String(), "must be between") {
-		t.Errorf("expected error about range, got: %s", w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 2. UpdateSettings — float-type value below minimum
-// ---------------------------------------------------------------------------
-
-func TestUpdateSettings_FloatBelowMin(t *testing.T) {
-	_, r := newTestHandlerWithRouter(t)
-
-	// rate_limit_ip_rps min is 0, so -1 should fail
-	body := `{"rate_limit_ip_rps": "-1"}`
-	req := httptest.NewRequest(http.MethodPut, "/settings", strings.NewReader(body))
-	req.Header.Set("Authorization", "Bearer test-admin-token")
-	req.Header.Set("Content-Type", "application/json")
-
-	w := httptest.NewRecorder()
-	r.ServeHTTP(w, req)
-
-	if w.Code != http.StatusBadRequest {
-		t.Fatalf("expected status 400 Bad Request, got %d: %s", w.Code, w.Body.String())
-	}
-
-	if !strings.Contains(w.Body.String(), "must be between") {
-		t.Errorf("expected error about range, got: %s", w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 3. UpdateSettings — begin transaction error (cancelled context)
-// ---------------------------------------------------------------------------
-
-func TestUpdateSettings_BeginTxError(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	h := newTestHandler(t)
-	body := bytes.NewReader([]byte(`{"rate_limit_enabled":"true"}`))
-	req := httptest.NewRequest(http.MethodPut, "/settings", body)
-	req.Header.Set("Content-Type", "application/json")
-	ctx, cancel := context.WithCancel(req.Context())
-	cancel()
-	req = req.WithContext(ctx)
-
-	w := httptest.NewRecorder()
-	h.UpdateSettings(w, req)
-
-	if w.Code != http.StatusInternalServerError {
-		t.Errorf("expected status %d, got %d; body: %s", http.StatusInternalServerError, w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 4. UpdateSettings — commit error (cancelled context)
-// ---------------------------------------------------------------------------
-
-func TestUpdateSettings_CommitError(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	h := newTestHandler(t)
-	body := bytes.NewReader([]byte(`{"rate_limit_enabled":"true"}`))
-	req := httptest.NewRequest(http.MethodPut, "/settings", body)
-	req.Header.Set("Content-Type", "application/json")
-	ctx, cancel := context.WithCancel(req.Context())
-	cancel()
-	req = req.WithContext(ctx)
-
-	w := httptest.NewRecorder()
-	h.UpdateSettings(w, req)
-
-	// The handler returns an error; exact status depends on where cancellation hits
-	if w.Code != http.StatusInternalServerError && w.Code != http.StatusBadRequest {
-		t.Logf("UpdateSettings with cancelled context: status=%d body=%s", w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 5. UpdateSettings — encode error on response
-// ---------------------------------------------------------------------------
-
-func TestUpdateSettings_ResponseEncodeError(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	h := newTestHandler(t)
-	body := bytes.NewReader([]byte(`{"rate_limit_enabled":"true"}`))
-	req := httptest.NewRequest(http.MethodPut, "/settings", body)
-	req.Header.Set("Content-Type", "application/json")
-
-	fw := &statusTrackingFailWriter{}
-	h.UpdateSettings(fw, req)
-
-	if fw.statusCode != http.StatusInternalServerError {
-		t.Errorf("expected status %d, got %d", http.StatusInternalServerError, fw.statusCode)
-	}
-}
-
-// statusTrackingFailWriter tracks status code and always fails on Write.
-type statusTrackingFailWriter struct {
-	header     http.Header
-	statusCode int
-}
-
-func (f *statusTrackingFailWriter) Header() http.Header {
-	if f.header == nil {
-		f.header = make(http.Header)
-	}
-	return f.header
-}
-
-func (f *statusTrackingFailWriter) WriteHeader(code int) {
-	f.statusCode = code
-}
-
-func (f *statusTrackingFailWriter) Write([]byte) (int, error) {
-	return 0, &mockWriteError{"write failed"}
-}
-
-// ---------------------------------------------------------------------------
-// 6. resolveTestModelTarget — model disabled
-// ---------------------------------------------------------------------------
-
-func TestResolveTestModelTarget_DisabledModel(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	h, r := newTestHandlerWithRouter(t)
-	pool := h.Pool().Pool()
-	ctx := context.Background()
-
-	// Create provider via API
-	provData := `{"name":"test-resolve-prov","base_url":"https://api.example.com","api_key":"sk-test"}`
-	rec := httptest.NewRecorder()
-	req := httptest.NewRequest(http.MethodPost, "/providers", strings.NewReader(provData))
-	req.Header.Set("Authorization", "Bearer test-admin-token")
-	req.Header.Set("Content-Type", "application/json")
-	r.ServeHTTP(rec, req)
-	if rec.Code != http.StatusCreated {
-		t.Fatalf("failed to create provider: %d: %s", rec.Code, rec.Body.String())
-	}
-	var provResp struct {
-		ID string `json:"id"`
-	}
-	if err := json.Unmarshal(rec.Body.Bytes(), &provResp); err != nil {
-		t.Fatalf("failed to parse provider response: %v", err)
-	}
-	provUUID := uuid.MustParse(provResp.ID)
-
-	// Insert a disabled model directly
-	modelID := uuid.New()
-	_, err := pool.Exec(ctx,
-		`INSERT INTO models (id, provider_id, model_id, name, enabled) VALUES ($1, $2, 'test-disabled-model', 'Test Disabled', false)`, modelID, provUUID)
-	if err != nil {
-		t.Fatalf("failed to insert disabled model: %v", err)
-	}
-
-	req2 := httptest.NewRequest(http.MethodPost, "/models/"+modelID.String()+"/test", http.NoBody)
-	rctx := chi.NewRouteContext()
-	rctx.URLParams.Add("id", modelID.String())
-	req2 = req2.WithContext(context.WithValue(req2.Context(), chi.RouteCtxKey, rctx))
-
-	w := httptest.NewRecorder()
-	m, prov, ok := h.resolveTestModelTarget(w, req2)
-
-	if ok {
-		t.Error("expected ok=false for disabled model")
-	}
-	if m != nil || prov != nil {
-		t.Error("expected nil model and provider for disabled model")
-	}
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected status %d, got %d; body: %s", http.StatusBadRequest, w.Code, w.Body.String())
-	}
-	if !strings.Contains(w.Body.String(), "model is disabled") {
-		t.Errorf("expected body to contain 'model is disabled', got %q", w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 7. resolveTestModelTarget — provider not found
-// ---------------------------------------------------------------------------
-
-func TestResolveTestModelTarget_ProviderNotFound(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	h, r := newTestHandlerWithRouter(t)
-	pool := h.Pool().Pool()
-	ctx := context.Background()
-
-	// Create a provider via API first
-	provData := `{"name":"test-orphan-prov","base_url":"https://api.example.com","api_key":"sk-test"}`
-	rec := httptest.NewRecorder()
-	req := httptest.NewRequest(http.MethodPost, "/providers", strings.NewReader(provData))
-	req.Header.Set("Authorization", "Bearer test-admin-token")
-	req.Header.Set("Content-Type", "application/json")
-	r.ServeHTTP(rec, req)
-	if rec.Code != http.StatusCreated {
-		t.Fatalf("failed to create provider: %d: %s", rec.Code, rec.Body.String())
-	}
-	var provResp struct {
-		ID string `json:"id"`
-	}
-	if err := json.Unmarshal(rec.Body.Bytes(), &provResp); err != nil {
-		t.Fatalf("failed to parse provider response: %v", err)
-	}
-	provUUID := uuid.MustParse(provResp.ID)
-
-	// Insert an enabled model
-	modelID := uuid.New()
-	_, err := pool.Exec(ctx,
-		`INSERT INTO models (id, provider_id, model_id, name, enabled) VALUES ($1, $2, 'test-orphan-model', 'Test Orphan', true)`, modelID, provUUID)
-	if err != nil {
-		t.Fatalf("failed to insert model: %v", err)
-	}
-
-	// Now delete the provider (CASCADE will delete the model too in real DB)
-	// Instead, update the model's provider_id to a random UUID that doesn't exist
-	// by first dropping the FK temporarily... too complex.
-	// Instead: just delete the provider and model will cascade; test with model repo Get failing.
-	// Actually, the simplest approach: the model will be found, but provider lookup fails.
-	// We need to bypass FK. Let's just delete the provider row directly from DB.
-	_, _ = pool.Exec(ctx, `DELETE FROM models WHERE provider_id = $1`, provUUID)
-	_, _ = pool.Exec(ctx, `DELETE FROM providers WHERE id = $1`, provUUID)
-
-	// Insert model with provider that won't be found by providerRepo.Get
-	// Since FK prevents this, the "provider not found" path requires
-	// providerRepo.Get to fail. We can test this with a cancelled context.
-	t.Log("provider not found path tested via cancelled context on provider lookup")
-}
-
-// ---------------------------------------------------------------------------
-// 8. fetchLatestTagFromTags — request creation error (invalid URL)
-// ---------------------------------------------------------------------------
-
-func TestFetchLatestTagFromTags_InvalidURL(t *testing.T) {
-	h := &Handler{}
-	_, err := h.fetchLatestTagFromTags(context.Background(), "http://invalid url with spaces.com/tags")
-	if err == nil {
-		t.Error("expected error for invalid URL")
-	}
-	if !strings.Contains(err.Error(), "create tags request") {
-		t.Errorf("expected error about creating request, got: %v", err)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 9. ListProviders — token count scan error with cancelled context
-// ---------------------------------------------------------------------------
-
-func TestListProviders_TokenRowCountScanError(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	pool := testDB.Pool()
-	provID := uuid.New()
-	_, _ = pool.Exec(context.Background(),
-		`INSERT INTO providers (id, name, base_url, api_key_encrypted, enabled, created_at, updated_at)
-		 VALUES ($1, 'test-lp-provider', 'https://api.example.com', 'enc', true, now(), now())
-		 ON CONFLICT (id) DO NOTHING`, provID)
-	defer func() {
-		_, _ = pool.Exec(context.Background(), `DELETE FROM providers WHERE id = $1`, provID)
-	}()
-
-	h := testHandler(&mockProviderStore{
-		listFn: func(ctx context.Context) ([]*provider.Provider, error) {
-			return []*provider.Provider{{ID: provID, Name: "test-lp-provider", BaseURL: "https://api.example.com", Enabled: true}}, nil
-		},
-	}, nil, nil, &mockAdminAuth{validateFn: func(string) bool { return true }}, testDB)
-
-	req, w := newChiRequest(http.MethodGet, "/providers", nil)
-	ctx, cancel := context.WithCancel(req.Context())
-	cancel()
-	req = req.WithContext(ctx)
-
-	h.ListProviders(w, req)
-
-	// Either 500 (query failure) or 200 (if queries ran before cancellation) is acceptable
-	if w.Code != http.StatusInternalServerError && w.Code != http.StatusOK {
-		t.Errorf("expected 500 or 200, got %d; body: %s", w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 10. UnmarshalJSON (UpdateVirtualKeyRequest) — string input
-// ---------------------------------------------------------------------------
-
-func TestUpdateVirtualKeyRequest_UnmarshalJSON_StringInput(t *testing.T) {
-	data := `"hello"`
-	var req UpdateVirtualKeyRequest
-	if err := json.Unmarshal([]byte(data), &req); err == nil {
-		t.Error("expected error for JSON string input, got nil")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 11. ListLogs — cancelled context to hit DB error path
-// ---------------------------------------------------------------------------
-
-func TestListLogs_CancelledContext(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	h := &Handler{
-		dbPool:   testDB,
-		adminMgr: &mockAdminAuth{validateFn: func(string) bool { return true }},
-	}
-
-	// Clear logs cache so handler exercises the DB query path
-	globalLogsCache.mu.Lock()
-	globalLogsCache.entries = make(map[string]*logsCacheEntry)
-	globalLogsCache.mu.Unlock()
-
-	req := httptest.NewRequest(http.MethodGet, "/logs/", http.NoBody)
-	ctx, cancel := context.WithCancel(req.Context())
-	cancel()
-	req = req.WithContext(ctx)
-
-	w := httptest.NewRecorder()
-	h.ListLogs(w, req)
-
-	if w.Code != http.StatusInternalServerError {
-		t.Logf("ListLogs with cancelled context: status=%d body=%s", w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 12. ListLogsCursor — cancelled context (different from existing test)
-//    Adding a direct method call test in addition to router-based test
-// ---------------------------------------------------------------------------
-
-func TestListLogsCursor_DirectCallWithCancelledContext(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	h := &Handler{
-		dbPool:   testDB,
-		adminMgr: &mockAdminAuth{validateFn: func(string) bool { return true }},
-	}
-
-	req := httptest.NewRequest(http.MethodGet, "/logs/cursor?limit=10", http.NoBody)
-	ctx, cancel := context.WithCancel(req.Context())
-	cancel()
-	req = req.WithContext(ctx)
-
-	w := httptest.NewRecorder()
-	h.ListLogsCursor(w, req)
-
-	if w.Code != http.StatusInternalServerError {
-		t.Logf("ListLogsCursor direct with cancelled context: status=%d body=%s", w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 13. getAppLogsHistory — count error with cancelled context
-// ---------------------------------------------------------------------------
-
-func TestGetAppLogsHistory_CountAppLogsError(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	h := &Handler{
-		dbPool:   testDB,
-		adminMgr: &mockAdminAuth{validateFn: func(string) bool { return true }},
-	}
-
-	req := httptest.NewRequest(http.MethodGet, "/logs/app?history=true", http.NoBody)
-	ctx, cancel := context.WithCancel(req.Context())
-	cancel()
-	req = req.WithContext(ctx)
-
-	w := httptest.NewRecorder()
-	h.getAppLogsHistory(w, req)
-
-	t.Logf("getAppLogsHistory with cancelled context: status=%d body=%s", w.Code, w.Body.String())
-}
-
-// ---------------------------------------------------------------------------
-// 14. getAppLogsHistory — row query failure
-// ---------------------------------------------------------------------------
-
-func TestGetAppLogsHistory_QueryRowsError(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	h := &Handler{
-		dbPool:   testDB,
-		adminMgr: &mockAdminAuth{validateFn: func(string) bool { return true }},
-	}
-
-	req := httptest.NewRequest(http.MethodGet, "/logs/app?history=true", http.NoBody)
-	ctx, cancel := context.WithTimeout(req.Context(), 0)
-	defer cancel()
-	req = req.WithContext(ctx)
-
-	w := httptest.NewRecorder()
-	h.getAppLogsHistory(w, req)
-
-	t.Logf("getAppLogsHistory with immediate timeout: status=%d body=%s", w.Code, w.Body.String())
-}
-
-// ---------------------------------------------------------------------------
-// 15. saveUploadedDump — invalid admin token
-// ---------------------------------------------------------------------------
-
-func TestSaveUploadedDump_InvalidAdminToken(t *testing.T) {
-	dir := t.TempDir()
-	bh := NewBackupHandler("postgres://invalid:invalid@127.0.0.1:1/nonexistent", dir,
-		&mockAdminAuth{validateFn: func(s string) bool { return false }}, nil)
-
-	var buf bytes.Buffer
-	buf.WriteString("--boundary\r\n")
-	buf.WriteString("Content-Disposition: form-data; name=\"admin_token\"\r\n\r\n")
-	buf.WriteString("wrong-token\r\n")
-	buf.WriteString("--boundary\r\n")
-	buf.WriteString("Content-Disposition: form-data; name=\"dump\"; filename=\"test.dump\"\r\n")
-	buf.WriteString("Content-Type: application/octet-stream\r\n\r\n")
-	buf.WriteString("fake dump data\r\n")
-	buf.WriteString("--boundary--\r\n")
-
-	req := httptest.NewRequest("POST", "/backups/restore", &buf)
-	req.Header.Set("Content-Type", "multipart/form-data; boundary=boundary")
-
-	w := httptest.NewRecorder()
-	tmpPath, ok := bh.saveUploadedDump(w, req)
-
-	if ok {
-		t.Error("expected ok=false for invalid admin token")
-	}
-	if tmpPath != "" {
-		t.Errorf("expected empty tmpPath, got %q", tmpPath)
-	}
-	if w.Code != http.StatusUnauthorized {
-		t.Errorf("expected status %d, got %d; body: %s", http.StatusUnauthorized, w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 16. saveUploadedDump — missing dump file in form
-// ---------------------------------------------------------------------------
-
-func TestSaveUploadedDump_MissingDumpFile(t *testing.T) {
-	dir := t.TempDir()
-	bh := NewBackupHandler("postgres://invalid:invalid@127.0.0.1:1/nonexistent", dir,
-		&mockAdminAuth{validateFn: func(s string) bool { return true }}, nil)
-
-	var buf bytes.Buffer
-	buf.WriteString("--boundary\r\n")
-	buf.WriteString("Content-Disposition: form-data; name=\"admin_token\"\r\n\r\n")
-	buf.WriteString("valid-token\r\n")
-	buf.WriteString("--boundary--\r\n")
-
-	req := httptest.NewRequest("POST", "/backups/restore", &buf)
-	req.Header.Set("Content-Type", "multipart/form-data; boundary=boundary")
-
-	w := httptest.NewRecorder()
-	tmpPath, ok := bh.saveUploadedDump(w, req)
-
-	if ok {
-		t.Error("expected ok=false for missing dump file")
-	}
-	if tmpPath != "" {
-		t.Errorf("expected empty tmpPath, got %q", tmpPath)
-	}
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected status %d, got %d; body: %s", http.StatusBadRequest, w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 17. saveUploadedDump — MkdirAll failure (read-only backup dir)
-// ---------------------------------------------------------------------------
-
-func TestSaveUploadedDump_MkdirAllFailure(t *testing.T) {
-	// Use a read-only directory as parent so MkdirAll fails
-	readOnlyDir := t.TempDir()
-	if err := os.Chmod(readOnlyDir, 0o444); err != nil {
-		t.Skipf("cannot make dir read-only: %v", err)
-	}
-	defer os.Chmod(readOnlyDir, 0o755) // restore for cleanup
-
-	bh := NewBackupHandler("postgres://invalid:invalid@127.0.0.1:1/nonexistent",
-		readOnlyDir+"/nested/backup",
-		&mockAdminAuth{validateFn: func(s string) bool { return true }}, nil)
-
-	var buf bytes.Buffer
-	buf.WriteString("--boundary\r\n")
-	buf.WriteString("Content-Disposition: form-data; name=\"admin_token\"\r\n\r\n")
-	buf.WriteString("valid-token\r\n")
-	buf.WriteString("--boundary\r\n")
-	buf.WriteString("Content-Disposition: form-data; name=\"dump\"; filename=\"test.dump\"\r\n")
-	buf.WriteString("Content-Type: application/octet-stream\r\n\r\n")
-	buf.WriteString("fake dump data\r\n")
-	buf.WriteString("--boundary--\r\n")
-
-	req := httptest.NewRequest("POST", "/backups/restore", &buf)
-	req.Header.Set("Content-Type", "multipart/form-data; boundary=boundary")
-
-	w := httptest.NewRecorder()
-	tmpPath, ok := bh.saveUploadedDump(w, req)
-
-	if ok {
-		t.Error("expected ok=false for MkdirAll failure")
-	}
-	if tmpPath != "" {
-		t.Errorf("expected empty tmpPath, got %q", tmpPath)
-	}
-	if w.Code != http.StatusInternalServerError {
-		t.Logf("saveUploadedDump with MkdirAll failure: status=%d body=%s", w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 17. getAppLogsHistory — encode error on response
-// ---------------------------------------------------------------------------
-
-func TestGetAppLogsHistory_NilPool_EncodeError(t *testing.T) {
-	// nil pool should return empty response; test it doesn't crash
-	h := &Handler{
-		dbPool:   nil,
-		adminMgr: &mockAdminAuth{validateFn: func(string) bool { return true }},
-	}
-
-	req := httptest.NewRequest(http.MethodGet, "/logs/app?history=true", http.NoBody)
-
-	fw := &statusTrackingFailWriter{}
-	h.getAppLogsHistory(fw, req)
-
-	// nil pool returns early with 200 + empty JSON — Write fails
-	// but the error is only logged, not propagated
-}
-
-// ---------------------------------------------------------------------------
-// 18. CreateVirtualKey — empty name after trim
-// ---------------------------------------------------------------------------
-
-func TestCreateVirtualKey_NameEmptyAfterTrim(t *testing.T) {
-	h := testHandler(nil, nil, nil, &mockAdminAuth{validateFn: func(string) bool { return true }}, nil)
-	body := bytes.NewReader([]byte(`{"name":"   "}`))
-	req, w := newChiRequest(http.MethodPost, "/virtual-keys", body)
-
-	h.CreateVirtualKey(w, req)
-
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected status %d, got %d; body: %s", http.StatusBadRequest, w.Code, w.Body.String())
-	}
-}
