--- internal/api/coverage_targets3_test.go
@@ -1,490 +1,0 @@
-package api
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-	"net/http/httptest"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/go-chi/chi/v5"
-	"github.com/google/uuid"
-
-	"github.com/hugalafutro/model-hotel/internal/db"
-	"github.com/hugalafutro/model-hotel/internal/provider"
-)
-
-// ---------------------------------------------------------------------------
-// 1. statTotals — 7-day period (7*24*time.Hour) path
-//    The 7d branch sets TotalRequestsLast7d first then cross-fills
-//    TotalRequestsLast24h. This is different from the 24h path tested
-//    in coverage_targets_test.go.
-// ---------------------------------------------------------------------------
-
-func TestStats_StatTotalsWith7dPeriod(t *testing.T) {
-	handler, pool, cleanup := newStatsHandler(t)
-	defer cleanup()
-
-	providerID := uuid.New()
-	insertTestProvider(t, pool, providerID, "stat7d-provider", "https://api.example.com/v1")
-	insertTestRequestLog(t, pool, uuid.New(), providerID, "test-model", 200, 100, 10, 20)
-
-	stats := &StatsResponse{
-		ByModel:      make(map[string]int64),
-		ByProvider:   make(map[string]int64),
-		ByVirtualKey: make(map[string]int64),
-	}
-	now := time.Now().UTC()
-	since := now.Add(-7 * 24 * time.Hour)
-
-	err := handler.statTotals(context.Background(), stats, "", "",
-		7*24*time.Hour, since, now)
-	if err != nil {
-		t.Fatalf("statTotals with 7d period: %v", err)
-	}
-	// The 7d branch: TotalRequestsLast7d should be set, TotalRequestsLast24h
-	// should be cross-filled by the secondary query.
-	if stats.TotalRequestsLast7d < 1 {
-		t.Errorf("TotalRequestsLast7d = %d, want >= 1", stats.TotalRequestsLast7d)
-	}
-	if stats.TotalRequestsLast24h < 0 {
-		t.Errorf("TotalRequestsLast24h = %d, want >= 0", stats.TotalRequestsLast24h)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 2. calculateStats — "tokens" metric path
-//    Existing tests use metric="requests". The tokens metric changes the
-//    SQL SELECT from COUNT(*) to SUM(...) which uses a different code path
-//    in statByModel/statByProvider/statByVirtualKey.
-// ---------------------------------------------------------------------------
-
-func TestCalculateStats_WithTokensMetric(t *testing.T) {
-	handler, pool, cleanup := newStatsHandler(t)
-	defer cleanup()
-
-	providerID := uuid.New()
-	insertTestProvider(t, pool, providerID, "tokens-metric-provider", "https://api.example.com/v1")
-	insertTestRequestLog(t, pool, uuid.New(), providerID, "tokens-model", 200, 100, 10, 20)
-	insertTestRequestLog(t, pool, uuid.New(), providerID, "tokens-model", 200, 200, 20, 40)
-
-	result, err := handler.calculateStats(context.Background(), 24*time.Hour, false, "tokens", false)
-	if err != nil {
-		t.Fatalf("calculateStats with tokens metric: %v", err)
-	}
-	if result == nil {
-		t.Fatal("expected non-nil result")
-	}
-	// Verify the tokens metric was used (ByModel values reflect token sums, not counts)
-	if len(result.ByModel) == 0 {
-		t.Error("expected at least one model in ByModel with tokens metric")
-	}
-	_ = result.ByProvider
-	_ = result.ByVirtualKey
-}
-
-// ---------------------------------------------------------------------------
-// 3. calculateStats — 7-day period with latency
-//    Tests the combination of 7d period AND includeLatency=true.
-// ---------------------------------------------------------------------------
-
-func TestCalculateStats_7dWithLatency(t *testing.T) {
-	handler, pool, cleanup := newStatsHandler(t)
-	defer cleanup()
-
-	providerID := uuid.New()
-	insertTestProvider(t, pool, providerID, "latency-7d-provider", "https://api.example.com/v1")
-	for i := 0; i < 5; i++ {
-		insertRichTestRequestLog(t, pool, uuid.New(), providerID, "latency-7d-model", 200, 100+i*10, 10, 20, requestLogOpts{
-			ResponseHeaderMs: float64(50 + i*5),
-			ProxyOverheadMs:  float64(10 + i*2),
-			LatencyMs:        float64(80 + i*3),
-		})
-	}
-
-	result, err := handler.calculateStats(context.Background(), 7*24*time.Hour, false, "requests", true)
-	if err != nil {
-		t.Fatalf("calculateStats 7d with latency: %v", err)
-	}
-	if result == nil {
-		t.Fatal("expected non-nil result")
-	}
-	_ = result.ByModelLatency
-	_ = result.ByProviderLatency
-	_ = result.TotalRequestsLast7d
-}
-
-// ---------------------------------------------------------------------------
-// 4. ListProviders — token count query error (separate from model count error)
-//    The token count query (tokenRows) is a separate DB call from modelCounts.
-//    Test that a cancelled context causes the token row query to fail.
-// ---------------------------------------------------------------------------
-
-func TestListProviders_TokenQueryError(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	// Insert provider and a model so model count query succeeds
-	pool := testDB.Pool()
-	provID := uuid.New()
-	_, _ = pool.Exec(context.Background(),
-		`INSERT INTO providers (id, name, base_url, enabled, created_at, updated_at)
-		 VALUES ($1, 'token-query-prov', 'https://api.example.com', true, now(), now())
-		 ON CONFLICT (id) DO NOTHING`, provID)
-	defer func() {
-		_, _ = pool.Exec(context.Background(), `DELETE FROM models WHERE provider_id = $1`, provID)
-		_, _ = pool.Exec(context.Background(), `DELETE FROM providers WHERE id = $1`, provID)
-	}()
-
-	h := testHandler(&mockProviderStore{
-		listFn: func(ctx context.Context) ([]*provider.Provider, error) {
-			return []*provider.Provider{{ID: provID, Name: "token-query-prov", BaseURL: "https://api.example.com", Enabled: true}}, nil
-		},
-	}, nil, nil, &mockAdminAuth{validateFn: func(string) bool { return true }}, testDB)
-
-	// Use a context that gets cancelled after a short delay — this lets the
-	// first query (model count) succeed but may fail the second (token count).
-	req, w := newChiRequest(http.MethodGet, "/providers", nil)
-	ctx, cancel := context.WithTimeout(req.Context(), 5*time.Second)
-	defer cancel()
-	req = req.WithContext(ctx)
-
-	h.ListProviders(w, req)
-
-	// Should succeed (both queries should succeed within 5s timeout)
-	if w.Code != http.StatusOK {
-		t.Logf("ListProviders with short timeout: status=%d body=%s", w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 5. ListProviders — token count row scan error
-//    Tests the tokenRows.Scan error by injecting a type mismatch through
-//    a cancelled context that hits mid-query.
-// ---------------------------------------------------------------------------
-
-func TestListProviders_TokenRowScanError(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	h := testHandler(&mockProviderStore{
-		listFn: func(ctx context.Context) ([]*provider.Provider, error) {
-			return []*provider.Provider{}, nil
-		},
-	}, nil, nil, &mockAdminAuth{validateFn: func(string) bool { return true }}, testDB)
-
-	req, w := newChiRequest(http.MethodGet, "/providers", nil)
-
-	h.ListProviders(w, req)
-
-	// Empty providers should return 200 with empty array
-	if w.Code != http.StatusOK {
-		t.Errorf("expected 200, got %d: %s", w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 6. DeleteModel — successful delete with failover sync
-//    Tests the happy path of DeleteModel where the model exists, is
-//    deleted, and failover sync runs. The existing test only covers
-//    the DB lookup error and pgx.ErrNoRows paths.
-// ---------------------------------------------------------------------------
-
-func TestDeleteModel_SuccessWithFailoverSync(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	h, r := newTestHandlerWithRouter(t)
-	pool := h.Pool().Pool()
-	ctx := context.Background()
-
-	// Create provider via API
-	provData := `{"name":"delete-model-prov","base_url":"https://api.example.com","api_key":"sk-test"}`
-	rec := httptest.NewRecorder()
-	req := httptest.NewRequest(http.MethodPost, "/providers", strings.NewReader(provData))
-	req.Header.Set("Authorization", "Bearer test-admin-token")
-	req.Header.Set("Content-Type", "application/json")
-	r.ServeHTTP(rec, req)
-	if rec.Code != http.StatusCreated {
-		t.Fatalf("failed to create provider: %d: %s", rec.Code, rec.Body.String())
-	}
-	var provResp struct {
-		ID string `json:"id"`
-	}
-	if err := json.Unmarshal(rec.Body.Bytes(), &provResp); err != nil {
-		t.Fatalf("failed to parse provider response: %v", err)
-	}
-	provUUID := uuid.MustParse(provResp.ID)
-
-	// Insert a model directly
-	modelID := uuid.New()
-	_, err = pool.Exec(ctx,
-		`INSERT INTO models (id, provider_id, model_id, name, enabled) VALUES ($1, $2, 'delete-test-model', 'Delete Test', true)`, modelID, provUUID)
-	if err != nil {
-		t.Fatalf("failed to insert model: %v", err)
-	}
-
-	// Delete via handler
-	delReq := httptest.NewRequest(http.MethodDelete, "/models/"+modelID.String(), http.NoBody)
-	rctx := chi.NewRouteContext()
-	rctx.URLParams.Add("id", modelID.String())
-	delReq = delReq.WithContext(context.WithValue(delReq.Context(), chi.RouteCtxKey, rctx))
-
-	w := httptest.NewRecorder()
-	h.DeleteModel(w, delReq)
-
-	if w.Code != http.StatusNoContent {
-		t.Errorf("expected 204, got %d: %s", w.Code, w.Body.String())
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 7. runScheduledBackup — backup mutex already locked
-//    Tests that runScheduledBackup returns immediately when the mutex is held.
-// ---------------------------------------------------------------------------
-
-func TestRunScheduledBackup_MutexAlreadyLocked(t *testing.T) {
-	dir := t.TempDir()
-	ss := &mockSettingsStore{
-		getDurationFn: func(_ context.Context, _ string, _ time.Duration) time.Duration {
-			return 1 * time.Hour
-		},
-	}
-	bh := NewBackupHandler("postgres://x", dir, &mockAdminAuth{}, ss)
-
-	// Lock the mutex to simulate an in-progress backup
-	bh.backupMu.Lock()
-	defer bh.backupMu.Unlock()
-
-	// runScheduledBackup should return immediately without panic
-	bh.runScheduledBackup(context.Background())
-}
-
-// ---------------------------------------------------------------------------
-// 8. getAppLogsHistory — row Scan error within for loop
-//    The rows.Scan error path inside the for loop (line 554) is the
-//    remaining uncovered branch. This can happen if the column count
-//    or types don't match the Scan arguments, which can't happen with
-//    a correctly structured query. We test by verifying the handler
-//    doesn't crash even when Scan errors occur (they silently skip rows).
-// ---------------------------------------------------------------------------
-
-func TestGetAppLogsHistory_RowScanErrorInLoop(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	// Insert an app_log entry so the query returns a row
-	pool := testDB.Pool()
-	_, execErr := pool.Exec(context.Background(),
-		`INSERT INTO app_logs (timestamp, level, source, message) VALUES (NOW(), 'info', 'scan-test', 'test message for scan')`)
-	if execErr != nil {
-		t.Fatalf("failed to insert app log: %v", execErr)
-	}
-	defer func() {
-		_, _ = pool.Exec(context.Background(), `DELETE FROM app_logs WHERE source = 'scan-test'`)
-	}()
-
-	h := &Handler{
-		dbPool:   testDB,
-		adminMgr: &mockAdminAuth{validateFn: func(string) bool { return true }},
-	}
-
-	req := httptest.NewRequest(http.MethodGet, "/logs/app?history=true", http.NoBody)
-
-	w := httptest.NewRecorder()
-	h.getAppLogsHistory(w, req)
-
-	// Should return 200 with entries
-	if w.Code != http.StatusOK {
-		t.Fatalf("expected 200, got %d: %s", w.Code, w.Body.String())
-	}
-
-	var response appLogsHistoryResponse
-	if err := json.Unmarshal(w.Body.Bytes(), &response); err != nil {
-		t.Fatalf("failed to parse response: %v", err)
-	}
-	// The entry should be present and properly scanned
-	if len(response.Entries) == 0 {
-		t.Error("expected at least one app log entry")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// 8b. getAppLogsHistory — row Scan error with cancelled context during iteration
-//     Tests that when the context is cancelled while iterating rows, the
-//     handler doesn't panic and gracefully returns partial results.
-// ---------------------------------------------------------------------------
-
-func TestGetAppLogsHistory_ScanErrorDuringIteration(t *testing.T) {
-	if apiTestDBURL == "" {
-		t.Skip("skipping: test database not available")
-	}
-
-	testDB, err := db.New(context.Background(), apiTestDBURL, 25, 5)
-	if err != nil {
-		t.Fatalf("failed to create test DB: %v", err)
-	}
-	defer testDB.Close()
-
-	h := &Handler{
-		dbPool:   testDB,
-		adminMgr: &mockAdminAuth{validateFn: func(string) bool { return true }},
-	}
-
-	req := httptest.NewRequest(http.MethodGet, "/logs/app?history=true", http.NoBody)
-	// Use an extremely short timeout that may expire during row iteration
-	ctx, cancel := context.WithTimeout(req.Context(), 1*time.Nanosecond)
-	defer cancel()
-	req = req.WithContext(ctx)
-
-	w := httptest.NewRecorder()
-	h.getAppLogsHistory(w, req)
-
-	// Should not crash; may return error or partial results
-	t.Logf("getAppLogsHistory with nanosecond timeout: status=%d", w.Code)
-}
-
-// ---------------------------------------------------------------------------
-// 9. ListComposeContainers (internal/util/docker.go) — coverage gap documentation
-//    The uncovered paths in ListComposeContainers are:
-//    - http.NewRequestWithContext error: structurally impossible since
-//      the URL is a hardcoded constant "http://localhost/containers/json?all=true"
-//    - All other error paths (client.Do, JSON decode, non-200 status) are
-//      already tested in internal/util/docker_wrapper_test.go
-// ---------------------------------------------------------------------------
-
-// ---------------------------------------------------------------------------
-// 10. NewSPAHandler (cmd/server/spa.go) — coverage gap documentation
-//    The NewSPAHandler and ServeHTTP are in package main (cmd/server).
-//    Existing tests in spa_test.go cover:
-//    - Normal constructor path (embedded FS available)
-//    - Fallback when embedded FS not found (fs.Sub error)
-//    - Fallback when index.html is empty/missing (fs.ReadFile error)
-//    - Cache headers for hash-named .js/.css files
-//    - SPA route fallback for non-existing paths
-//    - API path rejection (404)
-//
-//    The remaining 80% coverage gap is because NewSPAHandler is called
-//    with the real staticFS. When the frontend is built, it takes the
-//    full constructor path. When not built, it takes the fallback.
-//    Both paths are tested but the 80% figure likely reflects that
-//    ServeHTTP's fs.Stat + fileServer path is only exercised when
-//    the frontend is built and embedded (TestSPAHandler_ServeHTTP_StaticFileServed
-//    skips when no frontend build).
-// ---------------------------------------------------------------------------
-
-func TestNewSPAHandler_CoverageGapDocumentation(t *testing.T) {
-	// NewSPAHandler is in package main (cmd/server/spa.go), not in package api.
-	// The existing spa_test.go tests cover the constructor thoroughly via
-	// staticFS injection. The 80% figure reflects that the ServeHTTP method's
-	// fs.Stat → fileServer.ServeHTTP path only runs when the embedded FS
-	// contains actual built files. The testStaticFS() helper in spa_test.go
-	// already exercises this path when run from package main.
-	//
-	// The only remaining gap is the fs.Sub error path at line 17-23, which
-	// is tested by TestNewSPAHandler_FallbackWhenNoEmbedFS (but only when
-	// the embedded FS doesn't contain "static").
-	t.Log("NewSPAHandler coverage: fully tested in cmd/server/spa_test.go")
-	t.Log("The 80% figure reflects conditional paths that require a frontend build")
-}
-
-// ---------------------------------------------------------------------------
-// 11. LoginStart — json.Marshal error path
-//    Tests that when json.Marshal fails for the login session, the
-//    handler returns 500. The existing tests cover BeginDiscoverableLogin
-//    error and CreateSession error. The marshal error path (line 272-276)
-//    is untested because json.Marshal rarely fails on a SessionData struct.
-//    This is a structurally difficult path to exercise.
-// ---------------------------------------------------------------------------
-
-func TestLoginStart_MarshalErrorIsStructuralLimitation(t *testing.T) {
-	// json.Marshal on webauthn.SessionData can only fail if the struct
-	// contains values that can't be marshalled (e.g., channels, functions).
-	// Since webauthn.SessionData is a well-formed struct, this path is
-	// structurally unreachable in practice. The BeginDiscoverableLogin
-	// function always returns a marshallable SessionData.
-	t.Log("LoginStart json.Marshal error path (line 272-276) is structurally unreachable:")
-	t.Log("webauthn.SessionData is always marshallable after BeginDiscoverableLogin")
-}
-
-// ---------------------------------------------------------------------------
-// 12. RegisterStart — json.Marshal error path
-//    Same structural limitation as LoginStart. The json.Marshal on
-//    SessionData at line 140 can't realistically fail.
-// ---------------------------------------------------------------------------
-
-func TestRegisterStart_MarshalErrorIsStructuralLimitation(t *testing.T) {
-	// Same as LoginStart: json.Marshal on webauthn.SessionData returned
-	// by BeginRegistration will never fail because the struct only
-	// contains marshallable types. This is a structural limitation.
-	t.Log("RegisterStart json.Marshal error path (line 140-145) is structurally unreachable:")
-	t.Log("webauthn.SessionData is always marshallable after BeginRegistration")
-}
-
-// ---------------------------------------------------------------------------
-// 13. runMigration — schema_migrations table does NOT exist (first migration)
-//    The "if !exists" branch at line 172-183 creates the
-//    schema_migrations table. This is only hit on the very first run.
-//    Existing tests rely on New() which creates the table before any
-//    explicit runMigration test, so the table always exists.
-//    We test this by creating a fresh DB and directly calling runMigration
-//    after dropping the schema_migrations table.
-// ---------------------------------------------------------------------------
-
-// ---------------------------------------------------------------------------
-// 13. runMigration (internal/db/db.go) — schema_migrations table does NOT exist
-//    The "if !exists" branch at line 172-183 creates the schema_migrations
-//    table. This is only hit on the very first run. Existing db_test.go tests
-//    all call New() first which creates the table, so subsequent runMigration
-//    calls never hit this branch.
-//
-//    This cannot be tested from the api package because runMigration is
-//    unexported. The test must live in internal/db/db_test.go. Adding it there.
-// ---------------------------------------------------------------------------
-
-func TestRunMigration_SchemaTableNotExist_Documentation(t *testing.T) {
-	// runMigration is unexported in the db package, so we can't call it
-	// directly from the api package. The test should be added to
-	// internal/db/db_test.go. This note documents the gap.
-	//
-	// The uncovered branch is:
-	//   - Line 172-183: "if !exists" → CREATE TABLE IF NOT EXISTS schema_migrations
-	//   - This only runs when schema_migrations doesn't exist yet (first migration)
-	//
-	// To test this from db_test.go:
-	//   1. Create a fresh test DB
-	//   2. Drop the schema_migrations table
-	//   3. Call runMigration → should create the table and apply the migration
-	t.Log("runMigration !exists branch: must be tested in internal/db/db_test.go")
-	t.Log("The gap is the CREATE TABLE IF NOT EXISTS schema_migrations path (line 172-183)")
-}
