+++ internal/ratelimit/ip_limiter.go
@@ -23,10 +23,6 @@ const (
 	defaultIPBurst = 60
 )
 
-// cleanupInterval is the ticker interval for the background cleanup goroutine.
-// Overridable in tests for faster execution.
-var cleanupInterval = 5 * time.Minute
-
 // ipEntry tracks a single IP address's rate limiter.
 type ipEntry struct {
 	limiter  *rate.Limiter
@@ -184,7 +180,7 @@ func (l *IPLimiter) writeHeaders(w http.ResponseWriter, lim *rate.Limiter, retry
 }
 
 func (l *IPLimiter) cleanupLoop() {
-	ticker := time.NewTicker(cleanupInterval)
+	ticker := time.NewTicker(5 * time.Minute)
 	defer ticker.Stop()
 	for {
 		select {
