+++ internal/ratelimit/ip_limiter_test.go
@@ -1195,50 +1195,3 @@ func TestIPLimiter_CleanupLoop_TickerPathRemovesStaleEntries(t *testing.T) {
 // TestIPLimiter_CleanupRemovesStale and TestIPLimiter_CleanupLoop_TickerPathRemovesStaleEntries.
 // Only the select routing (ticker.C path) is untested; the actual cleanup
 // logic is fully covered. This is a structural limitation.
-
-// TestIPLimiter_CleanupLoop_TickerActuallyFires verifies that the cleanupLoop
-// actually invokes cleanup() when the ticker fires. By temporarily reducing
-// the cleanupInterval to a very short duration, we can observe the ticker
-// branch executing within a reasonable test timeout.
-func TestIPLimiter_CleanupLoop_TickerActuallyFires(t *testing.T) {
-	// Override the cleanup interval to make the ticker fire quickly
-	origInterval := cleanupInterval
-	cleanupInterval = 50 * time.Millisecond
-	defer func() { cleanupInterval = origInterval }()
-
-	lim := NewIPLimiter(10, 20, nil, nil)
-	defer lim.Stop()
-
-	// Insert a stale IP entry (last used 15 minutes ago)
-	lim.mu.Lock()
-	lim.limiters["ticker-actual-stale"] = &ipEntry{
-		limiter:  rate.NewLimiter(10, 20),
-		rps:      10,
-		burst:    20,
-		lastUsed: time.Now().Add(-15 * time.Minute),
-	}
-	// And a fresh entry that should survive
-	lim.limiters["ticker-actual-fresh"] = &ipEntry{
-		limiter:  rate.NewLimiter(10, 20),
-		rps:      10,
-		burst:    20,
-		lastUsed: time.Now(),
-	}
-	lim.mu.Unlock()
-
-	// Wait for the ticker to fire at least once (50ms interval, wait 200ms)
-	time.Sleep(200 * time.Millisecond)
-
-	// Check that the stale entry was removed by the ticker-fired cleanup
-	lim.mu.Lock()
-	_, hasStale := lim.limiters["ticker-actual-stale"]
-	_, hasFresh := lim.limiters["ticker-actual-fresh"]
-	lim.mu.Unlock()
-
-	if hasStale {
-		t.Error("stale IP entry should have been removed by ticker-fired cleanup")
-	}
-	if !hasFresh {
-		t.Error("fresh IP entry should still be present after ticker-fired cleanup")
-	}
-}
