+++ internal/ratelimit/limiter_test.go
@@ -1127,58 +1127,3 @@ func TestCleanupLoop_ConcurrentStopAndTick(t *testing.T) {
 // and TestCleanupLoop_Integration. Only the select case routing itself
 // (ticker.C vs stopCh) is untested; the actual cleanup logic is fully covered.
 // This is a structural limitation, not a gap in test intent.
-
-// TestCleanupLoop_TickerActuallyFires verifies that the cleanupLoop
-// actually invokes cleanup() when the ticker fires. By temporarily reducing
-// the cleanupInterval to a very short duration, we can observe the ticker
-// branch executing within a reasonable test timeout.
-func TestCleanupLoop_TickerActuallyFires(t *testing.T) {
-	// Override the cleanup interval to make the ticker fire quickly
-	origInterval := cleanupInterval
-	cleanupInterval = 50 * time.Millisecond
-	defer func() { cleanupInterval = origInterval }()
-
-	lim := &Limiter{
-		limiters: make(map[string]*keyEntry),
-		settings: newStubSettings(),
-		stopCh:   make(chan struct{}),
-	}
-
-	// Insert a stale entry (last used 15 minutes ago)
-	lim.mu.Lock()
-	lim.limiters["ticker-fires-stale"] = &keyEntry{
-		limiter:  rate.NewLimiter(10, 20),
-		rps:      10,
-		burst:    20,
-		lastUsed: time.Now().Add(-15 * time.Minute),
-	}
-	// And a fresh entry
-	lim.limiters["ticker-fires-fresh"] = &keyEntry{
-		limiter:  rate.NewLimiter(10, 20),
-		rps:      10,
-		burst:    20,
-		lastUsed: time.Now(),
-	}
-	lim.mu.Unlock()
-
-	// Start the cleanupLoop goroutine with the short ticker interval
-	go lim.cleanupLoop()
-
-	// Wait for the ticker to fire at least once (50ms interval, wait 200ms)
-	time.Sleep(200 * time.Millisecond)
-
-	// Check that the stale entry was removed by the ticker-fired cleanup
-	lim.mu.Lock()
-	_, hasStale := lim.limiters["ticker-fires-stale"]
-	_, hasFresh := lim.limiters["ticker-fires-fresh"]
-	lim.mu.Unlock()
-
-	if hasStale {
-		t.Error("stale key entry should have been removed by ticker-fired cleanup")
-	}
-	if !hasFresh {
-		t.Error("fresh key entry should still be present after ticker-fired cleanup")
-	}
-
-	lim.Stop()
-}
