+++ internal/webauthn/webauthn_test.go
@@ -6,7 +6,6 @@ import (
 	"crypto/sha256"
 	"encoding/hex"
 	"errors"
-	"io"
 	"log"
 	"os"
 	"testing"
@@ -1687,77 +1686,6 @@ func TestDeleteCredential_DatabaseError(t *testing.T) {
 	}
 }
 
-// TestToWebAuthnCredential_AAGUIDMarshalError verifies that ToWebAuthnCredential
-// handles AAGUID MarshalBinary errors gracefully by falling back to a 16-byte
-// zero slice. Since uuid.UUID.MarshalBinary never returns an error for a valid
-// UUID, we use the aaguidMarshalBinary override to inject the error.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestToWebAuthnCredential_AAGUIDMarshalError(t *testing.T) {
-	origFn := aaguidMarshalBinary
-	aaguidMarshalBinary = func(_ uuid.UUID) ([]byte, error) {
-		return nil, errors.New("simulated AAGUID MarshalBinary error")
-	}
-	defer func() { aaguidMarshalBinary = origFn }()
-
-	record := &CredentialRecord{
-		ID:              []byte("aaguid-err-id"),
-		PublicKey:       []byte("pub-key"),
-		AttestationType: "none",
-		AAGUID:          uuid.New(),
-		Transport:       []string{"internal"},
-		FlagsByte:       0x01,
-		SignCount:       5,
-	}
-
-	cred := record.ToWebAuthnCredential()
-	if string(cred.ID) != "aaguid-err-id" {
-		t.Errorf("expected ID 'aaguid-err-id', got %q", string(cred.ID))
-	}
-	// When MarshalBinary fails, the fallback is make([]byte, 16) — all zeros
-	if len(cred.Authenticator.AAGUID) != 16 {
-		t.Errorf("expected 16-byte AAGUID fallback, got %d bytes", len(cred.Authenticator.AAGUID))
-	}
-	for _, b := range cred.Authenticator.AAGUID {
-		if b != 0 {
-			t.Errorf("expected zero AAGUID bytes on error, got %x", cred.Authenticator.AAGUID)
-			break
-		}
-	}
-	if cred.Authenticator.SignCount != 5 {
-		t.Errorf("expected SignCount=5, got %d", cred.Authenticator.SignCount)
-	}
-}
-
-// TestToWebAuthnCredential_AAGUIDMarshalError_SuccessPath verifies the normal
-// (non-error) path through the aaguidMarshalBinary override.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestToWebAuthnCredential_AAGUIDMarshalError_SuccessPath(t *testing.T) {
-	testAAGUID := uuid.New()
-	origFn := aaguidMarshalBinary
-	aaguidMarshalBinary = func(u uuid.UUID) ([]byte, error) {
-		return u.MarshalBinary()
-	}
-	defer func() { aaguidMarshalBinary = origFn }()
-
-	record := &CredentialRecord{
-		ID:              []byte("aaguid-success-id"),
-		PublicKey:       []byte("pub-key"),
-		AttestationType: "none",
-		AAGUID:          testAAGUID,
-		Transport:       []string{"internal"},
-		FlagsByte:       0x01,
-		SignCount:       5,
-	}
-
-	cred := record.ToWebAuthnCredential()
-	if string(cred.ID) != "aaguid-success-id" {
-		t.Errorf("expected ID 'aaguid-success-id', got %q", string(cred.ID))
-	}
-	if len(cred.Authenticator.AAGUID) != 16 {
-		t.Errorf("expected 16-byte AAGUID, got %d bytes", len(cred.Authenticator.AAGUID))
-	}
-}
-
 // TestToWebAuthnCredential_EmptyAttestation verifies that all attestation fields
 // are propagated correctly when they are empty slices (not nil).
 func TestToWebAuthnCredential_EmptyAttestation(t *testing.T) {
@@ -1847,183 +1775,6 @@ func TestDeleteCredential_ContextCancelledMidTx(t *testing.T) {
 	}
 }
 
-// TestDeleteCredential_SessionDeleteTxError verifies that DeleteCredential returns
-// an error when the session DELETE tx.Exec fails. This covers the error path in
-// deleteCredSessionDelete. We use the txSessionDeleteFn override to inject the
-// error since the real DB cannot deterministically fail a tx.Exec after Begin succeeds.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestDeleteCredential_SessionDeleteTxError(t *testing.T) {
-	repo := newTestRepo(t)
-
-	origFn := txSessionDeleteFn
-	txSessionDeleteFn = func() error {
-		return errors.New("simulated session DELETE error")
-	}
-	defer func() { txSessionDeleteFn = origFn }()
-
-	err := repo.DeleteCredential(context.Background(), []byte("any-id"))
-	if err == nil {
-		t.Error("expected error from session DELETE tx failure")
-	}
-}
-
-// TestDeleteCredential_CredentialDeleteTxError verifies that DeleteCredential returns
-// an error when the credential DELETE tx.Exec fails. This covers the error path in
-// deleteCredCredentialDelete. We use the txCredentialDeleteFn override.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestDeleteCredential_CredentialDeleteTxError(t *testing.T) {
-	repo := newTestRepo(t)
-
-	origFn := txCredentialDeleteFn
-	txCredentialDeleteFn = func() error {
-		return errors.New("simulated credential DELETE error")
-	}
-	defer func() { txCredentialDeleteFn = origFn }()
-
-	err := repo.DeleteCredential(context.Background(), []byte("any-id"))
-	if err == nil {
-		t.Error("expected error from credential DELETE tx failure")
-	}
-}
-
-// TestDeleteCredential_CommitError verifies that DeleteCredential returns an error
-// when tx.Commit fails. This exercises the deleteCredCommit override path.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestDeleteCredential_CommitError(t *testing.T) {
-	repo := newTestRepo(t)
-
-	// Store a credential first
-	cred := &CredentialRecord{
-		ID:                []byte("commit-err-cred-id"),
-		PublicKey:         []byte("fake-public-key"),
-		AttestationType:   "none",
-		AttestationFormat: "packed",
-		Transport:         []string{"internal"},
-		FlagsByte:         0x41,
-		SignCount:         0,
-		AAGUID:            uuid.Nil,
-	}
-	if err := repo.StoreCredential(context.Background(), cred); err != nil {
-		t.Fatalf("StoreCredential: %v", err)
-	}
-
-	origCommitFn := txCommitFn
-	txCommitFn = func() error { return errors.New("simulated Commit error") }
-	defer func() { txCommitFn = origCommitFn }()
-
-	err := repo.DeleteCredential(context.Background(), cred.ID)
-	if err == nil {
-		t.Error("expected error from tx.Commit failure")
-	}
-
-	// Credential should still exist since commit failed
-	found, findErr := repo.GetCredentialByID(context.Background(), cred.ID)
-	if findErr != nil {
-		t.Fatalf("GetCredentialByID after failed commit: %v", findErr)
-	}
-	if string(found.ID) != "commit-err-cred-id" {
-		t.Errorf("credential should still exist after failed commit")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// CreateAuthToken error paths (structurally impossible without overrides)
-// ---------------------------------------------------------------------------
-
-// TestCreateAuthToken_RandReaderError verifies the io.ReadFull error path in
-// CreateAuthToken by overriding the rand reader with one that always fails.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestCreateAuthToken_RandReaderError(t *testing.T) {
-	repo := newTestRepo(t)
-	mgr := NewSessionManager(repo)
-
-	origReader := randReader
-	origReadFull := ioReadFull
-	randReader = &errorReader{}
-	ioReadFull = io.ReadFull
-	defer func() {
-		randReader = origReader
-		ioReadFull = origReadFull
-	}()
-
-	_, err := mgr.CreateAuthToken(context.Background(), []byte("admin"), nil)
-	if err == nil {
-		t.Error("expected error when rand.Reader fails in CreateAuthToken")
-	}
-}
-
-// errorReader is an io.Reader that always returns an error.
-type errorReader struct{}
-
-func (e *errorReader) Read(_ []byte) (int, error) {
-	return 0, errors.New("simulated rand.Reader error")
-}
-
-// TestCreateAuthToken_UUIDNewRandomError verifies the uuid.NewRandom error path
-// in CreateAuthToken by overriding the uuidNewRandom variable.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestCreateAuthToken_UUIDNewRandomError(t *testing.T) {
-	repo := newTestRepo(t)
-	mgr := NewSessionManager(repo)
-
-	origFn := uuidNewRandom
-	uuidNewRandom = func() (uuid.UUID, error) {
-		return uuid.Nil, errors.New("simulated uuid.NewRandom error")
-	}
-	defer func() { uuidNewRandom = origFn }()
-
-	_, err := mgr.CreateAuthToken(context.Background(), []byte("admin"), nil)
-	if err == nil {
-		t.Error("expected error when uuid.NewRandom fails in CreateAuthToken")
-	}
-}
-
-// TestCreateAuthToken_GenerateChallengeError verifies the generateChallenge error
-// path in CreateAuthToken by overriding the ioReadFull/randReader variables so
-// that generateChallenge's internal ReadFull call fails.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestCreateAuthToken_GenerateChallengeError(t *testing.T) {
-	repo := newTestRepo(t)
-	mgr := NewSessionManager(repo)
-
-	// Override ioReadFull so that the FIRST call (for tokenBytes in CreateAuthToken)
-	// succeeds, but the SECOND call (inside generateChallenge) fails.
-	callCount := 0
-	origReadFull := ioReadFull
-	ioReadFull = func(r io.Reader, buf []byte) (int, error) {
-		callCount++
-		if callCount == 1 {
-			// First call: succeed (fills tokenBytes)
-			copy(buf, make([]byte, len(buf)))
-			return len(buf), nil
-		}
-		// Second call: fail (generateChallenge)
-		return 0, errors.New("simulated generateChallenge error")
-	}
-	defer func() { ioReadFull = origReadFull }()
-
-	_, err := mgr.CreateAuthToken(context.Background(), []byte("admin"), nil)
-	if err == nil {
-		t.Error("expected error when generateChallenge fails in CreateAuthToken")
-	}
-}
-
-// TestGenerateChallenge_ReadFullError verifies that generateChallenge returns an
-// error when io.ReadFull fails.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestGenerateChallenge_ReadFullError(t *testing.T) {
-	origReadFull := ioReadFull
-	ioReadFull = func(_ io.Reader, _ []byte) (int, error) {
-		return 0, errors.New("simulated ReadFull error")
-	}
-	defer func() { ioReadFull = origReadFull }()
-
-	_, err := generateChallenge(32)
-	if err == nil {
-		t.Error("expected error from generateChallenge when ReadFull fails")
-	}
-}
-
 // ---------------------------------------------------------------------------
 // RevokeAuthToken error paths — DeleteSession failure
 // ---------------------------------------------------------------------------
@@ -2079,149 +1830,5 @@ func TestRevokeAuthToken_DeleteSessionFailure(t *testing.T) {
 	// Token should still be valid via the working manager
 	if !mgr.Validate(ctx, token) {
 		t.Error("expected token to still be valid after failed revocation with closed pool")
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Validate edge cases
-// ---------------------------------------------------------------------------
-
-// TestValidate_TokenHashMismatch verifies the subtle.ConstantTimeCompare != 1
-// branch in Validate. Since the DB queryGetSessionByTokenHash uses
-// WHERE token_hash = $1 (guaranteeing the computed hash matches the stored one),
-// this defense-in-depth branch cannot be triggered via a real DB.
-// We use the getSessionByTokenHashFn override to inject a session whose
-// TokenHash differs from the computed hash.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestValidate_TokenHashMismatch(t *testing.T) {
-	repo := newTestRepo(t)
-	mgr := NewSessionManager(repo)
-
-	// Create an auth token session normally so it exists in the DB
-	ctx := context.Background()
-	token, err := mgr.CreateAuthToken(ctx, []byte("admin"), nil)
-	if err != nil {
-		t.Fatalf("CreateAuthToken: %v", err)
-	}
-
-	// Verify token validates normally
-	if !mgr.Validate(ctx, token) {
-		t.Fatal("expected token to validate before override")
-	}
-
-	// Override getSessionByTokenHashFn to return a session with a DIFFERENT TokenHash
-	wrongHash := "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-	origFn := getSessionByTokenHashFn
-	getSessionByTokenHashFn = func(_ context.Context, _ string) (*SessionRecord, error) {
-		return &SessionRecord{
-			ID:        uuid.New(),
-			Type:      "auth_token",
-			TokenHash: &wrongHash,
-			ExpiresAt: time.Now().Add(1 * time.Hour),
-		}, nil
-	}
-	defer func() { getSessionByTokenHashFn = origFn }()
-
-	// Validate should return false because ConstantTimeCompare fails
-	if mgr.Validate(ctx, token) {
-		t.Error("expected Validate to return false when token hash mismatches")
-	}
-}
-
-// TestValidate_NilTokenHashDirect verifies the nil TokenHash defense-in-depth
-// branch in Validate. Since GetSessionByTokenHash uses WHERE token_hash = $1
-// (which excludes NULL rows), this branch cannot be triggered via a real DB.
-// We use the getSessionByTokenHashFn override to inject a session with nil TokenHash.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestValidate_NilTokenHashDirect(t *testing.T) {
-	repo := newTestRepo(t)
-	mgr := NewSessionManager(repo)
-
-	// Create an auth token session normally so it exists in the DB
-	ctx := context.Background()
-	token, err := mgr.CreateAuthToken(ctx, []byte("admin"), nil)
-	if err != nil {
-		t.Fatalf("CreateAuthToken: %v", err)
-	}
-
-	// Verify token validates normally
-	if !mgr.Validate(ctx, token) {
-		t.Fatal("expected token to validate before override")
-	}
-
-	// Override getSessionByTokenHashFn to return a session with nil TokenHash
-	origFn := getSessionByTokenHashFn
-	getSessionByTokenHashFn = func(_ context.Context, _ string) (*SessionRecord, error) {
-		return &SessionRecord{
-			ID:        uuid.New(),
-			Type:      "auth_token",
-			TokenHash: nil, // nil TokenHash — defense-in-depth check
-			ExpiresAt: time.Now().Add(1 * time.Hour),
-		}, nil
-	}
-	defer func() { getSessionByTokenHashFn = origFn }()
-
-	// Validate should return false because TokenHash is nil
-	if mgr.Validate(ctx, token) {
-		t.Error("expected Validate to return false when TokenHash is nil")
-	}
-}
-
-// TestRevokeAuthToken_DeleteSessionFnError verifies that RevokeAuthToken returns
-// false when GetSessionByTokenHash succeeds but DeleteSession fails.
-// This covers the debuglog.Error branch in RevokeAuthToken.
-// We use both getSessionByTokenHashFn and deleteSessionFn overrides to exercise
-// the test-only branches and the DeleteSession failure path.
-// NOTE: This mutates package-level variables; do not use t.Parallel() in this test.
-func TestRevokeAuthToken_DeleteSessionFnError(t *testing.T) {
-	repo := newTestRepo(t)
-	ctx := context.Background()
-	mgr := NewSessionManager(repo)
-
-	// Create a token so the session exists in the DB (for the real repo lookup)
-	token, err := mgr.CreateAuthToken(ctx, []byte("admin"), nil)
-	if err != nil {
-		t.Fatalf("CreateAuthToken: %v", err)
-	}
-
-	// Verify the token is valid (session exists)
-	if !mgr.Validate(ctx, token) {
-		t.Fatal("expected token to be valid before revocation attempt")
-	}
-
-	// Compute the token hash to return a session via the override
-	hash := sha256.Sum256([]byte(token))
-	tokenHash := hex.EncodeToString(hash[:])
-
-	// Look up the real session ID for the override to return
-	session, err := repo.GetSessionByTokenHash(ctx, tokenHash)
-	if err != nil {
-		t.Fatalf("GetSessionByTokenHash: %v", err)
-	}
-
-	// Override getSessionByTokenHashFn to return the real session (exercise override path)
-	origGetFn := getSessionByTokenHashFn
-	getSessionByTokenHashFn = func(_ context.Context, _ string) (*SessionRecord, error) {
-		return session, nil
-	}
-	defer func() { getSessionByTokenHashFn = origGetFn }()
-
-	// Override deleteSessionFn to simulate a DeleteSession failure
-	origDelFn := deleteSessionFn
-	deleteSessionFn = func(_ context.Context, _ uuid.UUID) error {
-		return errors.New("simulated DeleteSession error")
-	}
-	defer func() { deleteSessionFn = origDelFn }()
-
-	// RevokeAuthToken should return false because DeleteSession fails
-	if mgr.RevokeAuthToken(ctx, token) {
-		t.Error("expected RevokeAuthToken to return false when DeleteSession fails")
-	}
-
-	// Reset overrides and verify token still works via the real repo
-	getSessionByTokenHashFn = origGetFn
-	deleteSessionFn = origDelFn
-	if !mgr.Validate(ctx, token) {
-		t.Error("expected token to still be valid after failed DeleteSession")
 	}
 }
